+++ packages/config-loader/src/sources/ConfigSources.ts
@@ -107,7 +107,7 @@ export interface ConfigSourcesDefaultForTargetsOptions
  */
 export interface ConfigSourcesDefaultOptions extends BaseConfigSourcesOptions {
   argv?: string[];
-  env?: Record<string, string>;
+  env?: Record<string, string | undefined>;
 }
 
 /**