+++ packages/frontend-test-utils/src/app/createExtensionTester.tsx
@@ -19,6 +19,10 @@ import { MemoryRouter, Link } from 'react-router-dom';
 import { RenderResult, render } from '@testing-library/react';
 import { createSpecializedApp } from '@backstage/frontend-app-api';
 import {
+  AppNode,
+  AppTree,
+  Extension,
+  ExtensionDataRef,
   ExtensionDefinition,
   IconComponent,
   RouteRef,
@@ -30,12 +34,20 @@ import {
   createRouterExtension,
   useRouteRef,
 } from '@backstage/frontend-plugin-api';
-import { MockConfigApi } from '@backstage/test-utils';
+import { Config, ConfigReader } from '@backstage/config';
 import { JsonArray, JsonObject, JsonValue } from '@backstage/types';
 // eslint-disable-next-line @backstage/no-relative-monorepo-imports
 import { resolveExtensionDefinition } from '../../../frontend-plugin-api/src/wiring/resolveExtensionDefinition';
 // eslint-disable-next-line @backstage/no-relative-monorepo-imports
 import { toInternalExtensionDefinition } from '../../../frontend-plugin-api/src/wiring/createExtension';
+// eslint-disable-next-line @backstage/no-relative-monorepo-imports
+import { resolveAppTree } from '../../../frontend-app-api/src/tree/resolveAppTree';
+// eslint-disable-next-line @backstage/no-relative-monorepo-imports
+import { resolveAppNodeSpecs } from '../../../frontend-app-api/src/tree/resolveAppNodeSpecs';
+// eslint-disable-next-line @backstage/no-relative-monorepo-imports
+import { instantiateAppNodeTree } from '../../../frontend-app-api/src/tree/instantiateAppNodeTree';
+// eslint-disable-next-line @backstage/no-relative-monorepo-imports
+import { readAppExtensionsConfig } from '../../../frontend-app-api/src/tree/readAppExtensionsConfig';
 
 const NavItem = (props: {
   routeRef: RouteRef<undefined>;
@@ -43,10 +55,13 @@ const NavItem = (props: {
   icon: IconComponent;
 }) => {
   const { routeRef, title, icon: Icon } = props;
-  const to = useRouteRef(routeRef)();
+  const link = useRouteRef(routeRef);
+  if (!link) {
+    return null;
+  }
   return (
     <li>
-      <Link to={to}>
+      <Link to={link()}>
         <Icon /> {title}
       </Link>
     </li>
@@ -86,6 +101,35 @@ const TestAppNavExtension = createExtension({
 });
 
 /** @public */
+export class ExtensionQuery {
+  #node: AppNode;
+
+  constructor(node: AppNode) {
+    this.#node = node;
+  }
+
+  get node() {
+    return this.#node;
+  }
+
+  get instance() {
+    const instance = this.#node.instance;
+    if (!instance) {
+      throw new Error(
+        `Unable to access the instance of extension with ID '${
+          this.#node.spec.id
+        }'`,
+      );
+    }
+    return instance;
+  }
+
+  data<T>(ref: ExtensionDataRef<T>): T | undefined {
+    return this.instance.getData(ref);
+  }
+}
+
+/** @public */
 export class ExtensionTester {
   /** @internal */
   static forSubject<TConfig, TConfigInput>(
@@ -93,26 +137,54 @@ export class ExtensionTester {
     options?: { config?: TConfigInput },
   ): ExtensionTester {
     const tester = new ExtensionTester();
-    const { output, factory, ...rest } = toInternalExtensionDefinition(subject);
+    const internal = toInternalExtensionDefinition(subject);
+
     // attaching to app/routes to render as index route
-    const extension = createExtension({
-      ...rest,
-      attachTo: { id: 'app/routes', input: 'routes' },
-      output: {
-        ...output,
-        path: coreExtensionData.routePath,
-      },
-      factory: params => ({
-        ...factory(params),
-        path: '/',
-      }),
-    });
-    tester.add(extension, options as TConfigInput & {});
+    if (internal.version === 'v1') {
+      tester.add(
+        createExtension({
+          ...internal,
+          attachTo: { id: 'app/routes', input: 'routes' },
+          output: {
+            ...internal.output,
+            path: coreExtensionData.routePath,
+          },
+          factory: params => ({
+            ...internal.factory(params as any),
+            path: '/',
+          }),
+        }),
+        options as TConfigInput & {},
+      );
+    } else if (internal.version === 'v2') {
+      tester.add(
+        createExtension({
+          ...internal,
+          attachTo: { id: 'app/routes', input: 'routes' },
+          output: internal.output.find(
+            ref => ref.id === coreExtensionData.routePath.id,
+          )
+            ? internal.output
+            : [...internal.output, coreExtensionData.routePath],
+          factory: params => {
+            const parentOutput = Array.from(
+              internal.factory(params as any),
+            ).filter(val => val.id !== coreExtensionData.routePath.id);
+
+            return [...parentOutput, coreExtensionData.routePath('/')];
+          },
+        }),
+        options as TConfigInput & {},
+      );
+    }
     return tester;
   }
 
+  #tree?: AppTree;
+
   readonly #extensions = new Array<{
     id: string;
+    extension: Extension<any>;
     definition: ExtensionDefinition<any>;
     config?: JsonValue;
   }>();
@@ -121,6 +193,12 @@ export class ExtensionTester {
     extension: ExtensionDefinition<TConfig, TConfigInput>,
     options?: { config?: TConfigInput },
   ): ExtensionTester {
+    if (this.#tree) {
+      throw new Error(
+        'Cannot add more extensions accessing the extension tree',
+      );
+    }
+
     const { name, namespace } = extension;
 
     const definition = {
@@ -129,10 +207,11 @@ export class ExtensionTester {
       name: !namespace && !name ? 'test' : name,
     };
 
-    const { id } = resolveExtensionDefinition(definition);
+    const resolvedExtension = resolveExtensionDefinition(definition);
 
     this.#extensions.push({
-      id,
+      id: resolvedExtension.id,
+      extension: resolvedExtension,
       definition,
       config: options?.config as JsonValue,
     });
@@ -140,37 +219,41 @@ export class ExtensionTester {
     return this;
   }
 
+  data<T>(ref: ExtensionDataRef<T>): T | undefined {
+    const tree = this.#resolveTree();
+
+    return new ExtensionQuery(tree.root).data(ref);
+  }
+
+  query(id: string | ExtensionDefinition<any, any>): ExtensionQuery {
+    const tree = this.#resolveTree();
+
+    const actualId =
+      typeof id === 'string' ? id : resolveExtensionDefinition(id).id;
+
+    const node = tree.nodes.get(actualId);
+
+    if (!node) {
+      throw new Error(
+        `Extension with ID '${actualId}' not found, please make sure it's added to the tester.`,
+      );
+    } else if (!node.instance) {
+      throw new Error(
+        `Extension with ID '${actualId}' has not been instantiated, because it is not part of the test subject's extension tree.`,
+      );
+    }
+    return new ExtensionQuery(node);
+  }
+
   render(options?: { config?: JsonObject }): RenderResult {
     const { config = {} } = options ?? {};
 
-    const [subject, ...rest] = this.#extensions;
+    const [subject] = this.#extensions;
     if (!subject) {
       throw new Error(
         'No subject found. At least one extension should be added to the tester.',
       );
     }
-
-    const extensionsConfig: JsonArray = [
-      ...rest.map(extension => ({
-        [extension.id]: {
-          config: extension.config,
-        },
-      })),
-      {
-        [subject.id]: {
-          config: subject.config,
-          disabled: false,
-        },
-      },
-    ];
-
-    const finalConfig = {
-      ...config,
-      app: {
-        ...(typeof config.app === 'object' ? config.app : undefined),
-        extensions: extensionsConfig,
-      },
-    };
 
     const app = createSpecializedApp({
       features: [
@@ -187,10 +270,68 @@ export class ExtensionTester {
           ],
         }),
       ],
-      config: new MockConfigApi(finalConfig),
+      config: this.#getConfig(config),
     });
 
     return render(app.createRoot());
+  }
+
+  #resolveTree() {
+    if (this.#tree) {
+      return this.#tree;
+    }
+
+    const [subject] = this.#extensions;
+    if (!subject) {
+      throw new Error(
+        'No subject found. At least one extension should be added to the tester.',
+      );
+    }
+
+    const tree = resolveAppTree(
+      subject.id,
+      resolveAppNodeSpecs({
+        features: [],
+        builtinExtensions: this.#extensions.map(_ => _.extension),
+        parameters: readAppExtensionsConfig(this.#getConfig()),
+      }),
+    );
+
+    instantiateAppNodeTree(tree.root);
+
+    this.#tree = tree;
+
+    return tree;
+  }
+
+  #getConfig(additionalConfig?: JsonObject): Config {
+    const [subject, ...rest] = this.#extensions;
+
+    const extensionsConfig: JsonArray = [
+      ...rest.map(extension => ({
+        [extension.id]: {
+          config: extension.config,
+        },
+      })),
+      {
+        [subject.id]: {
+          config: subject.config,
+          disabled: false,
+        },
+      },
+    ];
+
+    return ConfigReader.fromConfigs([
+      { context: 'render-config', data: additionalConfig ?? {} },
+      {
+        context: 'test',
+        data: {
+          app: {
+            extensions: extensionsConfig,
+          },
+        },
+      },
+    ]);
   }
 }