+++ plugins/auth-backend-module-oidc-provider/config.d.ts
@@ -31,6 +31,12 @@ export interface Config {
           tokenSignedResponseAlg?: string;
           additionalScopes?: string | string[];
           prompt?: string;
+          signIn?: {
+            resolvers: Array<
+              | { resolver: 'emailLocalPartMatchingUserEntityName' }
+              | { resolver: 'emailMatchingUserEntityProfileEmail' }
+            >;
+          };
         };
       };
     };