+++ plugins/catalog-backend-module-gitlab/src/providers/config.test.ts
@@ -60,6 +60,7 @@ describe('config', () => {
         allowInherited: false,
         schedule: undefined,
         skipForkedRepos: false,
+        excludeRepos: [],
         restrictUsersToGroup: false,
       }),
     );
@@ -99,6 +100,7 @@ describe('config', () => {
         allowInherited: false,
         schedule: undefined,
         skipForkedRepos: false,
+        excludeRepos: [],
         restrictUsersToGroup: false,
       }),
     );
@@ -139,7 +141,50 @@ describe('config', () => {
         allowInherited: false,
         schedule: undefined,
         restrictUsersToGroup: false,
+        excludeRepos: [],
         skipForkedRepos: true,
+      }),
+    );
+  });
+
+  it('valid config with excludeRepos', () => {
+    const config = new ConfigReader({
+      catalog: {
+        providers: {
+          gitlab: {
+            test: {
+              group: 'group',
+              host: 'host',
+              branch: 'not-master',
+              fallbackBranch: 'main',
+              entityFilename: 'custom-file.yaml',
+              skipForkedRepos: false,
+              excludeRepos: ['foo/bar', 'quz/qux'],
+            },
+          },
+        },
+      },
+    });
+
+    const result = readGitlabConfigs(config);
+    expect(result).toHaveLength(1);
+    result.forEach(r =>
+      expect(r).toStrictEqual({
+        id: 'test',
+        group: 'group',
+        branch: 'not-master',
+        fallbackBranch: 'main',
+        host: 'host',
+        catalogFile: 'custom-file.yaml',
+        projectPattern: /[\s\S]*/,
+        groupPattern: /[\s\S]*/,
+        userPattern: /[\s\S]*/,
+        orgEnabled: false,
+        allowInherited: false,
+        schedule: undefined,
+        restrictUsersToGroup: false,
+        skipForkedRepos: false,
+        excludeRepos: ['foo/bar', 'quz/qux'],
       }),
     );
   });
@@ -181,6 +226,7 @@ describe('config', () => {
         allowInherited: false,
         skipForkedRepos: false,
         restrictUsersToGroup: false,
+        excludeRepos: [],
         schedule: {
           frequency: { minutes: 30 },
           timeout: {