+++ plugins/scaffolder/src/components/OngoingTask/OngoingTask.tsx
@@ -175,7 +175,9 @@ export const OngoingTask = (props: {
   return (
     <Page themeId="website">
       <Header
-        pageTitleOverride={`Run of ${templateName}`}
+        pageTitleOverride={
+          templateName ? `Run of ${templateName}` : `Scaffolder Run`
+        }
         title={
           <div>
             Run of <code>{templateName}</code>