+++ plugins/auth-backend-module-aws-alb-provider/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @backstage/plugin-auth-backend-module-aws-alb-provider
 
+## 0.1.15-next.2
+
+### Patch Changes
+
+- 93095ee: Make sure node-fetch is version 2.7.0 or greater
+- Updated dependencies
+  - @backstage/backend-plugin-api@0.8.0-next.2
+  - @backstage/backend-common@0.23.4-next.2
+  - @backstage/plugin-auth-backend@0.22.10-next.2
+  - @backstage/plugin-auth-node@0.5.0-next.2
+  - @backstage/errors@1.2.4
+
+## 0.1.15-next.1
+
+### Patch Changes
+
+- c8f1cae: Add `signIn` to authentication provider configuration schema
+- Updated dependencies
+  - @backstage/backend-plugin-api@0.7.1-next.1
+  - @backstage/plugin-auth-backend@0.22.10-next.1
+  - @backstage/backend-common@0.23.4-next.1
+  - @backstage/errors@1.2.4
+  - @backstage/plugin-auth-node@0.4.18-next.1
+
 ## 0.1.15-next.0
 
 ### Patch Changes