+++ plugins/signals-backend/CHANGELOG.md
@@ -1,5 +1,32 @@
 # @backstage/plugin-signals-backend
 
+## 0.1.9-next.2
+
+### Patch Changes
+
+- 93095ee: Make sure node-fetch is version 2.7.0 or greater
+- Updated dependencies
+  - @backstage/backend-plugin-api@0.8.0-next.2
+  - @backstage/backend-common@0.23.4-next.2
+  - @backstage/plugin-auth-node@0.5.0-next.2
+  - @backstage/plugin-events-node@0.3.9-next.2
+  - @backstage/plugin-signals-node@0.1.9-next.2
+  - @backstage/config@1.2.0
+  - @backstage/types@1.1.1
+
+## 0.1.9-next.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/backend-plugin-api@0.7.1-next.1
+  - @backstage/backend-common@0.23.4-next.1
+  - @backstage/config@1.2.0
+  - @backstage/types@1.1.1
+  - @backstage/plugin-auth-node@0.4.18-next.1
+  - @backstage/plugin-events-node@0.3.9-next.1
+  - @backstage/plugin-signals-node@0.1.9-next.1
+
 ## 0.1.9-next.0
 
 ### Patch Changes